Prepend XML parser functions with xml_
This might be a bit odd when the parser is used directly, but it makes
much more sense when it's imported (which is generally what we'll do
when using it internally).

Also, this simplifies the API a bit. The search and find functions
accept options that are passed to the xpath building function. This
simplifies the public API:

* xml_parse
* xml_search
* xml_find
* xml_text
* xml_attr